# Fix attribute value truncation