--- a/sdk/trace/span_test.go
+++ b/sdk/trace/span_test.go
@@ -320,7 +320,7 @@ func TestTruncate(t *testing.T) {
 			t.Run(tt.name, func(t *testing.T) {
 				t.Parallel()
 
-				got := safeTruncate(g.input, g.limit)
+				got := truncate(g.limit, g.input)
 				assert.Equalf(
 					t, g.expected, got,
 					"input: %q([]rune%v))\ngot: %q([]rune%v)\nwant %q([]rune%v)",
@@ -340,7 +340,7 @@ func BenchmarkTruncate(b *testing.B) {
 			b.RunParallel(func(pb *testing.PB) {
 				var out string
 				for pb.Next() {
-					out = safeTruncate(input, limit)
+					out = truncate(limit, input)
 				}
 				_ = out
 			})