@pellared 2024-11-25
What about the value truncation in `sdk/metric` and `sdk/log`?