@MrAlias 2024-11-25
> What about the value truncation in `sdk/metric`[...]?

[From the specification](https://github.com/open-telemetry/opentelemetry-specification/blob/88bffeac48aa5eb37ac8044e931af63a0b55fe00/specification/common/README.md#exempt-entities):

> Attributes, which belong to Metrics, are exempt from the limits described above at this time, as discussed in Metrics Attribute Limits.