@MrAlias 2024-11-25
> What about the value truncation in [...] `sdk/log`?

I imagine they need to be truncated and limits need to be honored there.

This PR is scoped to address the existing bug in the already implemented trace signal.