@MrAlias 2024-11-25
> > What about the value truncation in [...] `sdk/log`?
> 
> I imagine they need to be truncated and limits need to be honored there.

Tracking in https://github.com/open-telemetry/opentelemetry-go/issues/6004